+++ RELEASES.md
@@ -3,7 +3,8 @@
 ## 0.9.2dev
 
 #### New features
-+ Tweaked `get_backend` to ignore `None` inputs (PR # 525)
++ Added support for [Nearest Brenier Potentials (SSNB)](http://proceedings.mlr.press/v108/paty20a/paty20a.pdf) (PR #526)
++ Tweaked `get_backend` to ignore `None` inputs (PR #525)
 + Callbacks for generalized conditional gradient in `ot.da.sinkhorn_l1l2_gl` are now vectorized to improve performance (PR #507)
 
 #### Closed issues